requirements.txt
selenium==4.30.0
pandas==2.2.3
google-cloud-bigquery==3.31.0
pandas-gbq==0.28.0
dotenv==0.9.9
